+++ src/main/java/com/nineleaps/expensemanagementproject/repository/ExpenseRepository.java
@@ -33,6 +33,6 @@ public interface ExpenseRepository extends JpaRepository<Expense, Long> {
 
 	List<Expense> findByIsReportedAndIsHidden(boolean b, boolean b1);
 
-	List<Expense> findByEmployeeAndAmountAndDateAndCategoryAndMerchantName(Employee employee, Long amount,
+	List<Expense> findByEmployeeAndAmountAndDateAndCategoryAndMerchantName(Employee employee, Float amount,
 			LocalDate date, Category category, String merchantName);
 }
